requirements.txt
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.7.14
charset-normalizer==3.4.2
et_xmlfile==2.0.0
feedparser==6.0.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
requests==2.32.4
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
